services/product-service/requirements.txt
django==5.0
django-filter==25.1
djangorestframework==3.15
psycopg2-binary==2.9.9
dj-database-url==2.1.0
djangorestframework-simplejwt==5.3.1
pika==1.3.2              # RabbitMQ client
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
django-cors-headers==4.3.1
requests==2.31.0         # For sync HTTP calls between services
gunicorn==21.2.0         # Production WSGI server
whitenoise==6.12.0      # For serving static files in production
